sourcefiles/requirements.txt
awscli==1.19.93
boto3==1.17.93
botocore==1.20.93
colorama==0.4.3
docutils==0.15.2
jmespath==0.10.0
pyasn1==0.4.8
python-dateutil==2.8.1
PyYAML==5.4.1
rsa==4.7.2
s3transfer==0.4.2
six==1.16.0
urllib3==1.26.5
